backport MR 4846 to main

See merge request weblogic-cloud/weblogic-kubernetes-operator!4853

+++ integration-tests/src/test/java/oracle/weblogic/kubernetes/ItDiagnosticsFailedCondition.java
@@ -61,6 +61,7 @@ import static oracle.weblogic.kubernetes.TestConstants.IMAGE_PULL_POLICY;
 import static oracle.weblogic.kubernetes.TestConstants.MII_BASIC_APP_NAME;
 import static oracle.weblogic.kubernetes.TestConstants.MII_BASIC_IMAGE_NAME;
 import static oracle.weblogic.kubernetes.TestConstants.MII_BASIC_IMAGE_TAG;
+import static oracle.weblogic.kubernetes.TestConstants.SKIP_CLEANUP;
 import static oracle.weblogic.kubernetes.TestConstants.WEBLOGIC_IMAGE_TO_USE_IN_SPEC;
 import static oracle.weblogic.kubernetes.actions.ActionConstants.MODEL_DIR;
 import static oracle.weblogic.kubernetes.actions.TestActions.deleteClusterCustomResource;
@@ -708,6 +709,7 @@ class ItDiagnosticsFailedCondition {
    * type: Completed, status: false
    */
   @DisabledIfEnvironmentVariable(named = "ARM", matches = "true")
+  @DisabledIfEnvironmentVariable(named = "OCNE", matches = "true")
   @Test
   @DisplayName("Test domain status condition with managed server boot failure.")
   void testMSBootFailureStatus() {
@@ -844,11 +846,13 @@ class ItDiagnosticsFailedCondition {
       if (!testPassed) {
         LoggingUtil.generateLog(this, ns);
       }
-      if (assertDoesNotThrow(() -> clusterExists(clusterResName, CLUSTER_VERSION, domainNamespace).call())) {
-        deleteClusterCustomResource(clusterResName, domainNamespace);
-      }
-      if (assertDoesNotThrow(() -> domainExists(domainName, DOMAIN_VERSION, domainNamespace).call())) {
-        deleteDomainResource(domainNamespace, domainName);
+      if (!SKIP_CLEANUP) {
+        if (assertDoesNotThrow(() -> clusterExists(clusterResName, CLUSTER_VERSION, domainNamespace).call())) {
+          deleteClusterCustomResource(clusterResName, domainNamespace);
+        }
+        if (assertDoesNotThrow(() -> domainExists(domainName, DOMAIN_VERSION, domainNamespace).call())) {
+          deleteDomainResource(domainNamespace, domainName);
+        }
       }
     }
   }
